pyproject.toml
[project]
name = "ask-mary"
version = "0.1.0"
description = "AI clinical trial scheduling agent"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.0.0",
    "cloud-sql-python-connector[asyncpg]>=1.14.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.11",
    "twilio>=9.10.1",
    "google-cloud-storage>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "interrogate>=1.7.0",
]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
fail-under = 90

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
